--- kubekarma/controlleroperator/handlers/npts_job_template.yaml
@@ -1,27 +1,0 @@
-apiVersion: batch/v1
-kind: CronJob
-metadata:
-  name: hello
-spec:
-  schedule: "* * * * *"
-  successfulJobsHistoryLimit: 2
-  failedJobsHistoryLimit: 2
-  jobTemplate:
-    spec:
-      template:
-        spec:
-          containers:
-            - name: worker
-              image: "localhost:5000/network-policy-test-worker:4c5b4ae-dirty"
-              imagePullPolicy: IfNotPresent
-              env:
-                - name: NPTS_CONTROLLER_OPERATOR_URL
-                  value: http://network-policy-tester-operator-headless-service.network-policy-tester-operator.svc.cluster.local:8000
-                - name: WORKER_TASK_ID
-                  value: ""
-                - name: WORKER_CONTROLLER_VERSION
-                  value: ""
-                - name: WORKER_TASK_EXECUTION_CONFIG
-                  value: |
-                    ---
-          restartPolicy: OnFailure
