+++ kubekarma/controlleroperator/kinds/networktestsuite/__init__.py
@@ -1,0 +1,1 @@
+API_PLURAL = 'networktestsuites'
