+++ kubekarma/controlleroperator/kinds/types.py
@@ -1,0 +1,16 @@
+from typing import Optional, TypedDict
+
+
+class TestSuiteStatusType(TypedDict):
+    lastExecutionTime: str
+    lastExecutionErrorTime: str
+    lastSucceededTime: str
+    testExecutionStatus: str
+    testCases: list[dict]
+
+
+class TestCaseStatusType(TypedDict):
+    name: str
+    status: str
+    executionTime: str
+    error: Optional[str]
