+++ kubekarma/shared/crd/genericcrd.py
@@ -1,7 +1,5 @@
 """A module to represent the generic CRD object status."""
-import dataclasses
 import enum
-from typing import Optional
 
 
 class CRDTestPhase(enum.Enum):
