+++ kubekarma/worker/networksuite/readme.md
@@ -1,0 +1,5 @@
+```yaml
+apiVersion: kubekarma.io/v1
+kind: NetworkTestSuite
+```
+In this package you will found the NetworkTestSuite assertions.
