+++ protos/kubekarma/controlleroperator/grpcsrv/pb2/health.proto
@@ -1,4 +1,8 @@
 // https://github.com/grpc/grpc/blob/master/doc/health-checking.md
+
+/*
+  This service is for kubernetes probes.
+**/
 syntax = "proto3";
 
 package grpc.health.v1;
