+++ tox.ini
@@ -10,7 +10,7 @@ profile = black
 # Disable some pydocstyle checks:
 # I100, I101 and I201: import order, we use pylint for it
 ignore = D104, D100, D102, D101, D106, D107, W503, W504, I100, I101, I201
-exclude = */migrations/* */engine/*
+exclude = */pb2/*
 max-line-length = 100
 min_python_version = 3.10
